PHP 7.x compatibility
