+++ .travis.yml
@@ -33,17 +33,17 @@ matrix:
     - php: 5.6
       env: SYMFONY_VERSION=3.2.*
   allow_failures:
-    - php: 7.0
-    - php: 7.1
     - php: nightly
 
 before_script:
-  - pecl -q install xhprof-beta && echo "extension=xhprof.so" >> `php --ini | grep "Loaded Configuration" | sed -e "s|.*:\s*||"`
+  - export PHPUNIT="./vendor/bin/phpunit"
+  - if [[ ! "$TRAVIS_PHP_VERSION" = 7.* ]]; then pecl -q install xhprof-beta && echo "extension=xhprof.so" >> `php --ini | grep "Loaded Configuration" | sed -e "s|.*:\s*||"`; fi;
   - composer selfupdate
   - if [ "$SYMFONY_VERSION" != "" ]; then composer require "symfony/symfony:${SYMFONY_VERSION}" --no-update; fi;
   - if [[ "$TRAVIS_PHP_VERSION" == "5.5" && "$SYMFONY_VERSION" == "" ]]; then composer require "symfony/asset" --no-update; fi;
   - if [[ "$TRAVIS_PHP_VERSION" == "5.6" && "$SYMFONY_VERSION" == "" ]]; then composer require "symfony/asset" --no-update; fi;
   - if [[ "$TRAVIS_PHP_VERSION" = 7.* && "$SYMFONY_VERSION" == "" ]]; then composer require "symfony/asset" "symfony/yaml" --no-update; fi;
+  - if [[ "$TRAVIS_PHP_VERSION" = 7.* ]]; then ./.travis/install-tideways.sh && export PHPUNIT="php -dextension=/tmp/php-profiler-extension-4.1.1/modules/tideways.so vendor/bin/phpunit"; fi;
   - composer update --prefer-dist --no-interaction $COMPOSER_FLAGS
 
-script: ./vendor/bin/phpunit
+script: $PHPUNIT
